--- a/samples/CommunityToolkit.Maui.Markup.Sample/HotReloadHandler.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/HotReloadHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-
 namespace CommunityToolkit.Maui.Markup.Sample;
 
 class HotReloadHandler : ICommunityToolkitHotReloadHandler