--- a/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/Android/MainActivity.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/Android/MainActivity.cs
@@ -1,6 +1,5 @@
 ﻿using Android.App;
 using Android.Content.PM;
-
 namespace CommunityToolkit.Maui.Markup.Sample;
 
 [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]