--- a/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/Android/MainApplication.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/Android/MainApplication.cs
@@ -1,6 +1,5 @@
 ﻿using Android.App;
 using Android.Runtime;
-
 namespace CommunityToolkit.Maui.Markup.Sample;
 
 [Application]