--- a/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/MacCatalyst/Program.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/MacCatalyst/Program.cs
@@ -1,5 +1,4 @@
 ﻿using UIKit;
-
 namespace CommunityToolkit.Maui.Markup.Sample;
 
 public class Program