--- a/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/iOS/AppDelegate.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/iOS/AppDelegate.cs
@@ -1,5 +1,4 @@
 ﻿using Foundation;
-
 namespace CommunityToolkit.Maui.Markup.Sample;
 
 [Register(nameof(AppDelegate))]