--- a/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/iOS/Program.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Platforms/iOS/Program.cs
@@ -1,5 +1,4 @@
 ﻿using UIKit;
-
 namespace CommunityToolkit.Maui.Markup.Sample;
 
 public class Program