--- a/samples/CommunityToolkit.Maui.Markup.Sample/Services/HackerNewsApiService.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Services/HackerNewsApiService.cs
@@ -1,6 +1,4 @@
-﻿using Polly;
-
-namespace CommunityToolkit.Maui.Markup.Sample.Services;
+﻿namespace CommunityToolkit.Maui.Markup.Sample.Services;
 
 class HackerNewsAPIService
 {