--- a/samples/CommunityToolkit.Maui.Markup.Sample/Services/IHackerNewsApi.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/Services/IHackerNewsApi.cs
@@ -1,5 +1,4 @@
 ﻿using Refit;
-
 namespace CommunityToolkit.Maui.Markup.Sample.Services;
 
 interface IHackerNewsApi