--- a/samples/CommunityToolkit.Maui.Markup.Sample/ViewModels/Base/BaseViewModel.cs
+++ b/samples/CommunityToolkit.Maui.Markup.Sample/ViewModels/Base/BaseViewModel.cs
@@ -1,5 +1,5 @@
 ﻿namespace CommunityToolkit.Maui.Markup.Sample.ViewModels.Base;
 
-abstract partial class BaseViewModel : ObservableObject
+abstract class BaseViewModel : ObservableObject
 {
 }