--- a/src/CommunityToolkit.Maui.Markup.SourceGenerators/IsExternalInit.cs
+++ b/src/CommunityToolkit.Maui.Markup.SourceGenerators/IsExternalInit.cs
@@ -41,8 +41,8 @@
 
 namespace System.Runtime.CompilerServices
 {
-	using System.Diagnostics;
-	using System.Diagnostics.CodeAnalysis;
+	using global::System.Diagnostics;
+	using global::System.Diagnostics.CodeAnalysis;
 	/// <summary>
 	///     Reserved to be used by the compiler for tracking metadata.
 	///     This class should not be used by developers in source code.