--- a/src/CommunityToolkit.Maui.Markup.UnitTests/Base/BaseMarkupTestFixture.cs
+++ b/src/CommunityToolkit.Maui.Markup.UnitTests/Base/BaseMarkupTestFixture.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using Microsoft.Maui.Controls;
-using NUnit.Framework;
-
+﻿using NUnit.Framework;
 namespace CommunityToolkit.Maui.Markup.UnitTests.Base;
 
 abstract class BaseMarkupTestFixture<TBindable> : BaseMarkupTestFixture where TBindable : BindableObject, new()