--- a/src/CommunityToolkit.Maui.Markup.UnitTests/Base/BaseTestFixture.cs
+++ b/src/CommunityToolkit.Maui.Markup.UnitTests/Base/BaseTestFixture.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Globalization;
+﻿using System.Globalization;
 using CommunityToolkit.Maui.Markup.UnitTests.Mocks;
-using Microsoft.Maui.Dispatching;
 using NUnit.Framework;
-
 namespace CommunityToolkit.Maui.Markup.UnitTests.Base;
 
 abstract class BaseTestFixture
@@ -14,8 +11,8 @@ abstract class BaseTestFixture
 	[SetUp]
 	public virtual void Setup()
 	{
-		defaultCulture = System.Threading.Thread.CurrentThread.CurrentCulture;
-		defaultUICulture = System.Threading.Thread.CurrentThread.CurrentUICulture;
+		defaultCulture = Thread.CurrentThread.CurrentCulture;
+		defaultUICulture = Thread.CurrentThread.CurrentUICulture;
 
 		DispatcherProvider.SetCurrent(new MockDispatcherProvider());
 	}
@@ -23,8 +20,8 @@ abstract class BaseTestFixture
 	[TearDown]
 	public virtual void TearDown()
 	{
-		System.Threading.Thread.CurrentThread.CurrentCulture = defaultCulture ?? throw new NullReferenceException();
-		System.Threading.Thread.CurrentThread.CurrentUICulture = defaultUICulture ?? throw new NullReferenceException();
+		Thread.CurrentThread.CurrentCulture = defaultCulture ?? throw new NullReferenceException();
+		Thread.CurrentThread.CurrentUICulture = defaultUICulture ?? throw new NullReferenceException();
 
 		DispatcherProvider.SetCurrent(null);
 	}