--- a/src/CommunityToolkit.Maui.Markup.UnitTests/BindableLayoutExtensionsTests.cs
+++ b/src/CommunityToolkit.Maui.Markup.UnitTests/BindableLayoutExtensionsTests.cs
@@ -1,6 +1,5 @@
 ﻿using CommunityToolkit.Maui.Markup.UnitTests.Base;
 using NUnit.Framework;
-
 namespace CommunityToolkit.Maui.Markup.UnitTests;
 
 [TestFixture]