--- a/src/CommunityToolkit.Maui.Markup.UnitTests/DynamicResourceHandlerTests.cs
+++ b/src/CommunityToolkit.Maui.Markup.UnitTests/DynamicResourceHandlerTests.cs
@@ -1,5 +1,4 @@
 ﻿using NUnit.Framework;
-
 namespace CommunityToolkit.Maui.Markup.UnitTests;
 
 public class DynamicResourceHandlerTests