--- a/src/CommunityToolkit.Maui.Markup.UnitTests/GridRowsColumns.cs
+++ b/src/CommunityToolkit.Maui.Markup.UnitTests/GridRowsColumns.cs
@@ -59,7 +59,7 @@ class GridRowsColumns : BaseMarkupTestFixture
 	{
 		Assert.Throws<ArgumentException>(
 			() => Rows.Define((Row.First, 8), (Row.Third, 8)),
-			$"Value of row name Third is not 1. " +
+			"Value of row name Third is not 1. " +
 			"Rows must be defined with enum names whose values form the sequence 0,1,2,...");
 	}
 
@@ -112,7 +112,7 @@ class GridRowsColumns : BaseMarkupTestFixture
 	{
 		Assert.Throws<ArgumentException>(
 			() => Columns.Define((Col.Second, 8), (Col.First, 8)),
-			$"Value of column name Second is not 0. " +
+			"Value of column name Second is not 0. " +
 			"Columns must be defined with enum names whose values form the sequence 0,1,2,...");
 	}