… Directory.Build.props
… src/CommunityToolkit.Maui.Markup.Benchmarks/Benchmarks/ExecuteBindingsBase.cs
… src/CommunityToolkit.Maui.Markup.Benchmarks/Benchmarks/ExecuteBindings_ViewModelToView.cs
… src/CommunityToolkit.Maui.Markup.Benchmarks/Benchmarks/ExecuteBindings_ViewToViewModel.cs
… src/CommunityToolkit.Maui.Markup.Benchmarks/Benchmarks/InitializeBindings.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/ItemsViewTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/LabelExtensionsLeftToRightTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/LabelExtensionsRightToLeftTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/LabelExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/Mocks/MockApplication.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/Mocks/MockResourcesProvider.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/ObjectExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/PaceholderExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/PaddingElementExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/SemanticPropertiesExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/StyleTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/TextAlignmentExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/TextTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/TypedBindingExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/UnitExpressionSearch.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/ViewExtensionsLeftToRightTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/ViewExtensionsRightToLeftTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/ViewExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup.UnitTests/VisualElementExtensionsTests.cs
… src/CommunityToolkit.Maui.Markup/AbsoluteLayoutExtensions.cs
… src/CommunityToolkit.Maui.Markup/AppBuilderExtensions.cs
… src/CommunityToolkit.Maui.Markup/BindableLayoutExtensions.cs
… src/CommunityToolkit.Maui.Markup/BindableObjectExtensions.cs
… src/CommunityToolkit.Maui.Markup/BindableObjectMultiBindExtensions.cs
… src/CommunityToolkit.Maui.Markup/CommunityToolkitMetadataUpdateHandler.cs