### Description of Change ###

This is a PR focuses on improving the benchmarks of `CommunityToolkit.Maui.Markup`.

The biggest change is creating a custom `TypedBinding` class, implementing allocation + performance improvements.