@brminnick 2024-09-06
The initial results look good! I'm seeing an improvement across the board on our benchmarks.

## Initialize Binding Benchmark

### Before

| Method                          | Mean       | Error     | StdDev    | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------- |-----------:|----------:|----------:|------:|--------:|-------:|-------:|----------:|------------:|
| InitializeDefaultBindings       |   9.528 μs | 0.1891 μs | 0.2024 μs |  1.00 |    0.03 | 0.0916 | 0.0763 |   2.53 KB |        1.00 |
| InitializeDefaultBindingsMarkup |   9.383 μs | 0.1803 μs | 0.2146 μs |  0.99 |    0.03 | 0.0916 | 0.0763 |   2.53 KB |        1.00 |
| InitializeTypedBindingsMarkup   | 175.076 μs | 3.0166 μs | 2.8217 μs | 18.38 |    0.47 |      - |      - |   11.4 KB |        4.50 |

### After 

| Method                          | Mean       | Error     | StdDev    | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------- |-----------:|----------:|----------:|------:|--------:|-------:|-------:|----------:|------------:|
| InitializeDefaultBindings       |   9.642 μs | 0.1383 μs | 0.1294 μs |  1.00 |    0.02 | 0.1373 | 0.1221 |   2.53 KB |        1.00 |
| InitializeDefaultBindingsMarkup |   9.564 μs | 0.1662 μs | 0.1474 μs |  0.99 |    0.02 | 0.1373 | 0.1221 |   2.53 KB |        1.00 |
| InitializeTypedBindingsMarkup   | 135.366 μs | 1.0701 μs | 1.0010 μs | 14.04 |    0.21 | 0.4883 | 0.2441 |  11.54 KB |        4.56 |

## Execute ViewModelToViewModel Binding Benchmarks

### Before 

| Method                                       | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| ExecuteDefaultBindings_ViewToViewModel       | 691.2 ns |  9.11 ns |  8.52 ns |  1.00 |    0.02 | 0.0057 |     160 B |        1.00 |
| ExecuteDefaultBindingsMarkup_ViewToViewModel | 701.8 ns |  8.93 ns |  8.35 ns |  1.02 |    0.02 | 0.0057 |     160 B |        1.00 |
| ExecuteTypedBindingsMarkup_ViewToViewModel   | 696.5 ns | 12.83 ns | 12.00 ns |  1.01 |    0.02 | 0.0057 |     160 B |        1.00 |

### After

| Method                                       | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| ExecuteDefaultBindings_ViewToViewModel       | 671.2 ns | 11.96 ns | 11.19 ns |  1.00 |    0.02 | 0.0076 |     160 B |        1.00 |
| ExecuteDefaultBindingsMarkup_ViewToViewModel | 676.1 ns | 13.06 ns | 12.21 ns |  1.01 |    0.02 | 0.0076 |     160 B |        1.00 |
| ExecuteTypedBindingsMarkup_ViewToViewModel   | 679.3 ns |  7.01 ns |  6.22 ns |  1.01 |    0.02 | 0.0076 |     160 B |        1.00 |


## Execute ViewModelToView Binding Benchmarks

### Before

| Method                                       | Mean     | Error    | StdDev   | Ratio | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|---------:|---------:|------:|----------:|------------:|
| ExecuteDefaultBindings_ViewModelToView       | 12.92 ns | 0.142 ns | 0.133 ns |  1.00 |         - |          NA |
| ExecuteDefaultBindingsMarkup_ViewModelToView | 11.85 ns | 0.089 ns | 0.079 ns |  0.92 |         - |          NA |
| ExecuteTypedBindingsMarkup_ViewModelToView   | 11.79 ns | 0.119 ns | 0.111 ns |  0.91 |         - |          NA |

### After 

| Method                                       | Mean      | Error     | StdDev    | Ratio | RatioSD | Allocated | Alloc Ratio |
|--------------------------------------------- |----------:|----------:|----------:|------:|--------:|----------:|------------:|
| ExecuteDefaultBindings_ViewModelToView       | 11.003 ns | 0.1422 ns | 0.1330 ns |  1.00 |    0.02 |         - |          NA |
| ExecuteDefaultBindingsMarkup_ViewModelToView |  9.972 ns | 0.1550 ns | 0.1374 ns |  0.91 |    0.02 |         - |          NA |
| ExecuteTypedBindingsMarkup_ViewModelToView   | 10.169 ns | 0.0882 ns | 0.0782 ns |  0.92 |    0.01 |         - |          NA |