@brminnick 2024-09-07
I'm ready to merge this PR as I'm seeing a marked improvement across the board on our benchmarks.

## Initialize Binding Benchmark

### Before

| Method                          | Mean       | Error     | StdDev    | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------- |-----------:|----------:|----------:|------:|--------:|-------:|-------:|----------:|------------:|
| InitializeDefaultBindings       |   9.528 μs | 0.1891 μs | 0.2024 μs |  1.00 |    0.03 | 0.0916 | 0.0763 |   2.53 KB |        1.00 |
| InitializeDefaultBindingsMarkup |   9.383 μs | 0.1803 μs | 0.2146 μs |  0.99 |    0.03 | 0.0916 | 0.0763 |   2.53 KB |        1.00 |
| InitializeTypedBindingsMarkup   | 175.076 μs | 3.0166 μs | 2.8217 μs | 18.38 |    0.47 |      - |      - |   11.4 KB |        4.50 |

### After 

| Method                          | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------- |----------:|----------:|----------:|------:|--------:|-------:|-------:|----------:|------------:|
| InitializeDefaultBindings       |  7.026 μs | 0.0728 μs | 0.0645 μs |  1.00 |    0.01 | 0.0992 | 0.0916 |   2.53 KB |        1.00 |
| InitializeDefaultBindingsMarkup |  7.025 μs | 0.0420 μs | 0.0372 μs |  1.00 |    0.01 | 0.0992 | 0.0916 |   2.53 KB |        1.00 |
| InitializeTypedBindingsMarkup   | 97.673 μs | 0.2341 μs | 0.1955 μs | 13.90 |    0.13 | 0.2441 |      - |  11.05 KB |        4.36 |

## Execute ViewModelToViewModel Binding Benchmarks

### Before 

| Method                                       | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| ExecuteDefaultBindings_ViewToViewModel       | 691.2 ns |  9.11 ns |  8.52 ns |  1.00 |    0.02 | 0.0057 |     160 B |        1.00 |
| ExecuteDefaultBindingsMarkup_ViewToViewModel | 701.8 ns |  8.93 ns |  8.35 ns |  1.02 |    0.02 | 0.0057 |     160 B |        1.00 |
| ExecuteTypedBindingsMarkup_ViewToViewModel   | 696.5 ns | 12.83 ns | 12.00 ns |  1.01 |    0.02 | 0.0057 |     160 B |        1.00 |

### After

| Method                                       | Mean     | Error   | StdDev  | Ratio | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
| ExecuteDefaultBindings_ViewToViewModel       | 558.4 ns | 0.85 ns | 0.79 ns |  1.00 | 0.0057 |     160 B |        1.00 |
| ExecuteDefaultBindingsMarkup_ViewToViewModel | 542.2 ns | 0.58 ns | 0.54 ns |  0.97 | 0.0057 |     160 B |        1.00 |
| ExecuteTypedBindingsMarkup_ViewToViewModel   | 566.0 ns | 0.39 ns | 0.34 ns |  1.01 | 0.0057 |     160 B |        1.00 |


## Execute ViewModelToView Binding Benchmarks

### Before

| Method                                       | Mean     | Error    | StdDev   | Ratio | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|---------:|---------:|------:|----------:|------------:|
| ExecuteDefaultBindings_ViewModelToView       | 12.92 ns | 0.142 ns | 0.133 ns |  1.00 |         - |          NA |
| ExecuteDefaultBindingsMarkup_ViewModelToView | 11.85 ns | 0.089 ns | 0.079 ns |  0.92 |         - |          NA |
| ExecuteTypedBindingsMarkup_ViewModelToView   | 11.79 ns | 0.119 ns | 0.111 ns |  0.91 |         - |          NA |

### After 

| Method                                       | Mean     | Error     | StdDev    | Ratio | Allocated | Alloc Ratio |
|--------------------------------------------- |---------:|----------:|----------:|------:|----------:|------------:|
| ExecuteDefaultBindings_ViewModelToView       | 8.103 ns | 0.0032 ns | 0.0028 ns |  1.00 |         - |          NA |
| ExecuteDefaultBindingsMarkup_ViewModelToView | 8.102 ns | 0.0033 ns | 0.0029 ns |  1.00 |         - |          NA |
| ExecuteTypedBindingsMarkup_ViewModelToView   | 8.098 ns | 0.0050 ns | 0.0042 ns |  1.00 |         - |          NA |